+++ marketplace/backend/api/v1/endpoints.js
@@ -70,6 +70,7 @@ export const Redemption = {
   prefix: '/redemption',
   get: '/:id',
   create: '/',
+  getRedemptionServices: '/services',
   getOutgoingRedemptionRequests: '/outgoing',
   getIncomingRedemptionRequests: '/incoming',
   close: '/close'
@@ -147,36 +148,14 @@ export const Order = {
   get: '/:address',
   getAll: '/',
   create: '/',
-  updateOrderStatus: '/update',
-  updateBuyerDetails: '/updateBuyerDetails',
-  updateSellerDetails: '/updateSellerDetails',
   payment: '/payment',
-  paymentSession: '/payment/session/:session_id/:sellersCommonName',
-  paymentIntent: '/payment/intent/:session_id/:sellersCommonName',
   userAddress: '/userAddress',
-  getAllUserAddress: '/userAddresses/user',
-  getAddressFromId: '/userAddress/:id',
-  createSaleOrder: '/sale',
+  getAllUserAddress: '/userAddresses/user/:redemptionService?',
   cancelSaleOrder: '/sale/cancel',
   checkSaleQuantity: '/saleQuantity',
   executeSale: '/closeSale',
   updateOrderComment: '/updateComment',
   export: '/exportOrders'
-}
-
-export const OrderLine = {
-  prefix: '/orderLine',
-  get: '/:address'
-}
-
-export const OrderLineItem = {
-  prefix: '/orderLineItem',
-  get: '/:orderLineId',
-  getAll: '/',
-  create: '/',
-  update: '/update',
-  audit: '/:address/:chainId/audit',
-  transferOwnership: '/transferOwnership',
 }
 
 export const Marketplace = {
@@ -192,10 +171,8 @@ export const Marketplace = {
 
 export const PaymentService = {
   prefix: '/payment',
-  stripeOnboarding: '/stripe/account',
-  stripeConnectStatus: '/stripe/account/status/:ownerCommonName',
-  stripeWebhook: '/stripe/webhook',
-  stripeWebhookConnect: '/stripe/webhook/connect',
+  getAll: '/',
+  getNotOnboarded: '/onboarding',
 }
 
 export const UserActivity = {